pyproject.toml
[tool.poetry]
name = "fun-with-models"
version = "0.1.0"
description = ""
authors = ["Samuel Moore <samuel.moore@disneystreaming.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.7.3"
email-validator = "^1.1.2"
python-slugify = "^4.0.1"
black = "^20.8b1"
pylama = "^7.7.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
